% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGenesFromCpGs.R
\name{GetGenesFromCpGs}
\alias{GetGenesFromCpGs}
\title{GetGenesFromCpGs}
\usage{
Give CpGs, get Genes (and such)
}
\arguments{
\item{sig.cpg}{CpG list}
}
\description{
GetGenesFromCpGs
}
\examples{
Result = GetGenesFromCpGs(sig.cpg = c("cg25871843","cg26405148"), array.type = "EPIC")
}
