% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCCx.R
\name{MCCx}
\alias{MCCx}
\title{MCCx}
\usage{
MCCx(data, lev = NULL, model = NULL, showCM = FALSE)
}
\arguments{
\item{data}{Input the dataset object (column 1 = OBS; column 2 = PRED).}

\item{lev}{NULL}

\item{model}{NULL}

\item{showCM}{Show output of confusion matrix & metrics as cat.}
}
\value{
Several metrics to indicate machine learning performance ("MCC","Accuracy","Recall","Precision","F1","CombinedScore")
}
\description{
This function required a dataset with column 1 as observations and column 2 with predictions. Multiclass!
}
\examples{
MCCx(data.frame(obs=c(1,1,1,2,2,2,3,3,3),pred=c(1,1,1,1,2,2,3,3,3)))
MCCx(data.frame(obs=c(rep(0,10),rep(1,10)),pred=c(rep(0,10),rep(1,10))))
MCCx(data.frame(obs=c(rep(0,10),rep(1,10)),pred=c(rep(1,10),rep(0,10))))
MCCx(data.frame(obs=c(rep(0,10),rep(1,10),rep(2,10)),pred=c(rep(0,10),rep(1,10),rep(2,10))))
}
