% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPCAproject.R
\name{PCAproject}
\alias{PCAproject}
\title{RRnorm}
\usage{
PCAproject()
}
\arguments{
\item{X}{Object that shoulb be normalized.}

\item{Scalar}{Object that shoulb be scaled by.}

\item{Offset}{Object that shoulb be Offseted by.}
}
\value{
Normalized object
}
\description{
This function normalizes dataset and offests/scales if needed.
}
\examples{
Object=seq(from = -1,to = 1,by = 0.1)^2
NormObject = RRnorm(Object, Scalar = 2, Offset = -1)
plot(NormObject, col = "red")
points(Object, col="blue")
}
