% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMCC.R
\name{pMCC}
\alias{pMCC}
\title{pMCC}
\usage{
pMCC(data, lev = NULL, model = NULL, showCM = FALSE, level1 = NA, s_TH = 0.5)
}
\arguments{
\item{data}{Input the dataset object (column 1 = OBS; column 2 = PRED).}

\item{lev}{NULL}

\item{model}{NULL}

\item{showCM}{Show output of confusion matrix & metrics as cat.}

\item{s_TH}{TH for setting classes form probabilities.}

\item{lev/level1}{The actual label/number of the first (to be compared to) 'class'. NEEDS FIXING @RRR}
}
\value{
Several metrics to indicate machine learning performance ("MCC","Accuracy","Recall","Precision","F1","CombinedScore")
}
\description{
This function required a dataset with column 1 as observations and column 2 with predictions. Two-class only!
}
\examples{
pMCC(data.frame(obs=c(1,1,1,2,2,2),pred=c(0.1,0.1,0.4999,0.5001,0.9,0.9)))
pMCC(data.frame(obs=c(1,1,1,2,2,2),pred=c(0.1,0.1,0.2,0.8,0.9,0.9)))
}
