% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPCAprojectv2.R
\name{ProjectTimePCA}
\alias{ProjectTimePCA}
\title{ProjectTimePCA}
\usage{
ProjectTimePCA(Data_t0 = NA, Data_t1 = NA, ClassColumnID = NA)
}
\arguments{
\item{Data_t0}{Input the dataset object from time-point 0 (row=sample; col=feature).}

\item{Data_t1}{Input the dataset object from time-point 1 (row=sample; col=feature).}

\item{ClassColumnID}{Index number of class column (eg. for iris dataset classcolumn ID is 5 (species)).}
}
\value{
A list of prcomp objects, distances and plots.
}
\description{
This function will creata a PCA of data at time 0, then data from time 2 will be projected in the same PCA. The biggest difference between groups is calculated thoughout all PCs. The PC with the highest difference is the PC of interest, therefore the weighted coefs are shown for that PC; indicative of the contibutors of change.
}
\examples{
# make PC plot
a=prcomp(iris[,-5])

# show plot
autoplot(a, data = iris, colour = 'Species')

# make modification within one feature (as time point 2)
irisMod = iris
irisMod[irisMod$Species=="virginica",c("Sepal.Length")] = irisMod[irisMod$Species=="setosa",c("Sepal.Length")] + 1

# make PC plot
b=prcomp(irisMod[,-5])

# show time point 2 plot
autoplot(b, data = irisMod, colour = 'Species')

# store as time-points
Data_t0 = iris
Data_t1 = irisMod

# run function
ProjectTimePCA(Data_t0 = Data_t0,Data_t1 = Data_t1,ClassColumnID = 5)
}
