% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcor.R
\name{Rcor}
\alias{Rcor}
\title{Rcor}
\usage{
Rcor(df = NA, th = 0.3)
}
\arguments{
\item{th}{threshold the correlations given a cutoff; will be shrunk to 0.}

\item{dataset}{Input the dataset object (data.frame).}
}
\value{
A correlation matrix; containing all columns able to be transformed to numeric.
}
\description{
This function requires a dataset with variables able to be transformed to numeric (not character).
}
\examples{
Rcor(mtcars)
}
