% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMostContibutingFeatures.R
\name{RMostContibutingFeatures}
\alias{RMostContibutingFeatures}
\title{RMostContibutingFeatures}
\usage{
RMostContibutingFeatures(
  PCA = NA,
  maxPCs = 10,
  Quantile_TH_range = 0.9,
  use_ABS = FALSE,
  plotfigure = TRUE
)
}
\arguments{
\item{PCA}{Input the PCA object (data.frame).}

\item{maxPCs}{(Default 10) The amount of PCs to be evaluated, but will select less if prcomp has less components.}

\item{Quantile_TH_range}{(Default 0.9) The quantile range to detemine the data to be selected in modelling the average increase.}

\item{use_ABS}{(Default FALSE) Use the signed (no ABS) loadings or unsigned (ABS) loadings.}

\item{plotfigure}{(Default TRUE) Plot the loadings in decreasing order, shows the model and 3x the sd of residual of model; then plots the features above threshold.}
}
\value{
A list of features driving the variance of selected components, with the relative variance explained in the main PC. Can be used to reduce data, select features related to biggest obseved variance.
}
\description{
This function requires a PCA object of a dataset.
}
\examples{
RMostContibutingFeatures(prcomp(mtcars))
}
