% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMostContibutingFeatures2.R
\name{RMostContibutingFeatures2}
\alias{RMostContibutingFeatures2}
\title{RMostContibutingFeatures2}
\usage{
This function requires a Dat object of a dataset, based on a vector of importances.
}
\arguments{
\item{Dat}{Input the Dat object (vector).}

\item{amountSDTH}{(Default 3) The amount SDs to be passed before considering the value to be relevant}

\item{Quantile_TH_range}{(Default 0.9) The quantile range to detemine the data to be selected in modelling the average increase.}

\item{plotfigure}{(Default TRUE) Plot the loadings in decreasing order, shows the model and 3x the sd of residual of model; then plots the features above threshold.}
}
\value{
The names of the most important features.
}
\description{
RMostContibutingFeatures2
}
\examples{
Dat = data.frame(Vals = c(rnorm(100),(1.0+(1:10)/10)^3)); rownames(Dat) = paste0("Feature",rownames(Dat))
Results = RMostContibutingFeatures2(Dat)
}
