% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRFE.R
\name{RRFE}
\alias{RRFE}
\title{RRFE}
\usage{
RRFE(
  dataset = NA,
  f_dataset_class_column_id = NA,
  s_AmountFeatures = NA,
  s_MinimalVariance = 0.5,
  s_MaxComponents = 50,
  s_KRepeats = 25,
  s_KmeansRepeat = 10,
  verbose = TRUE,
  ShowPlots = FALSE,
  s_GetLoadingDistance = FALSE,
  s_AngleFilter = FALSE,
  s_MakeFilteredObject = FALSE,
  s_scale = FALSE,
  s_ZNorm = FALSE,
  s_ShowLoadingArrows = TRUE
)
}
\arguments{
\item{dataset}{Input the dataset object (row=sample; col=feature; 'class' column advised).}

\item{f_dataset_class_column_id}{Index number of class column (eg. for iris dataset classcolumn ID is 5 (species)).}

\item{s_AmountFeatures}{Retrieves X amount of important features from PCA.}

\item{s_MinimalVariance}{Cumulative relative contribution of loading coefficients.}

\item{s_MaxComponents}{Maximum number of PCs to evaluate.}

\item{s_KRepeats}{Repeat k-means clustering s_KRepeats times.}

\item{s_KmeansRepeat}{Repeat k-means clustering SCORING s_KmeansRepeat times.}

\item{verbose}{More info during running.}

\item{ShowPlots}{Show the plots during running.}

\item{s_GetLoadingDistance}{Gets the distances between loading vectors and k-means centres (classes).}

\item{s_AngleFilter}{Use orthogonal loadings to filter the data on; selecting only features relevant to TWO groups.}

\item{s_MakeFilteredObject}{Create the filtered output, based on the important features.}

\item{s_scale}{Normalise the features by scaling (prcomp). Default to FALSE as some features might not be detected otherwise.}

\item{s_ZNorm}{Normalise the features by scaling (Z score scaling). Default to FALSE as some features might not be detected otherwise.}
}
\value{
A list of imporant features in a designed structure.
}
\description{
This function required a dataset with rows as samples and columns as features.
It should contain a "class" column to be able to detect classes. Feature-extraction
will be performed using PCA and seeded-kmeans. RRFEDefaults() can be used to set default settings.
}
\examples{
dataset = iris, f_dataset_class_column_id = which(colnames(iris)=="Species"), s_MinimalVariance = 0.5, s_MaxComponents = 50, s_KRepeats = 25, s_KmeansRepeat = 10, verbose = TRUE, ShowPlots = TRUE, s_MakeFilteredObject = TRUE
}
