% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRFilt.R
\name{RRFilt}
\alias{RRFilt}
\title{RRFilt}
\usage{
RRFilt(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
An indication of 'difference' in the given vector. Higher is better diff.
}
\description{
This function gives an indication how big the difference is in selected variable. When applying normally distributed features with no effect results in abs(x-mean(x))/sd(x), resulting near to 0. When applying normally distributed features with high effect the mean != median, thus abs(x-mean(x))/sd(x). This should result in higher than 0 number, higher the better difference in
}
\examples{
apply(iris[,-5],2,RRFilt)
}
