% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRlda.R
\name{RRlda}
\alias{RRlda}
\title{RRlda}
\usage{
RRlda(
  dataset = iris,
  f_dataset_class_column_id = NA,
  s_formula = "Class~.",
  minimumfeaturesPercentagePlaceholdertext = 0.2,
  plotPCAImage = TRUE,
  minimumAmountOfFeatures = 2
)
}
\arguments{
\item{dataset}{Input the dataset object.}

\item{f_dataset_class_column_id}{Column number of "Class" column. Detected automatically if this column is called "Class" (case sensitive).}

\item{s_formula}{The formula used in the LDA. Should be like "Class~." (class determined by remaining columns)}

\item{minimumfeaturesPercentagePlaceholdertext}{PLACEHOLDER; extracts % from ordered top IQR;}

\item{plotPCAImage}{Plot a PCA image from filtered dataset.}

\item{minimumAmountOfFeatures}{Minimum amount of resulted featured used to in PCA plot (and dataset filtering).}
}
\value{
A list of raw LDA, filtered andordered IQR values, filtered dataset.
}
\description{
This will extract the most contributing features from a dataset, based on the given formula. Internally computed by LDA, the IQR between classes is calculated and ordered by amplitude. Biggest amplitude holds the biggest difference between classes, thus should contibute to classification.
}
\examples{
RRlda(iris,5)
}
