% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRRemoveCor.R
\name{RRRemoveCor}
\alias{RRRemoveCor}
\title{RRRemoveCor}
\usage{
RRRemoveCor(dataset, TH = 0.7)
}
\arguments{
\item{dataset}{Input the dataset in dataframe format (row=sample, col=feature).}

\item{TH}{ABS Threshold correlation (defauls to 0.7)}
}
\value{
Dataframe with modified values for correlated features.
}
\description{
This function required a dataset with rows as samples and columns as features.
All features should be continues, not categorical or ordinal as the func cannot deal with those at this time.
The fuction will assess the pearson corrlation between all the given features, and identify the most correlated ones.
Then the corrlated features will be linearly modelled by each other, extracting fitted and residuals in the process.
Internally it selects the most likely model based on normally disributed residuals.
The concept is that corrlated variables share information, and the only unique information between them are the residuals.
This function will result a modified dataframe, with updated values for the correlated features.
}
\examples{
Updated_iris = RRRemoveCor(iris[,-5])
}
