% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRscan.R
\name{RRscan}
\alias{RRscan}
\title{RRscan}
\usage{
RRscan(
  data = data,
  color = 1,
  LogTransform = FALSE,
  ImageDir = "",
  PlotImages = TRUE,
  SaveImages = FALSE,
  FindOutliers = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{data}{Data object in dataframe with samples as rows and features as columns.}

\item{color}{A vector indicating the colors per sample.}

\item{LogTransform}{Transform data using log2(x+1).}

\item{ImageDir}{The directory location for images to be written to.}

\item{PlotImages}{Plot images in the plotter.}

\item{SaveImages}{Save the images to in the ImageDir}

\item{FindOutliers}{Print potential outliers based on 3 testes failed (hclust dist, mean+3*sd, mean(variance) + 3*sd(variance))}

\item{verbose}{Give more information.}
}
\value{
Multiple plots giving an overview of the data.
}
\description{
This function will generate information and images about inputted data, getting fundamental properties of entered data.
}
\examples{
RRscan(data = mtcars,verbose=TRUE)
}
