% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRZnorm.R
\name{Znorm}
\alias{Znorm}
\title{Znorm}
\usage{
Znorm(X, NormalizePerRow = TRUE)
}
\arguments{
\item{X}{Object that should be normalized.}

\item{NormalizePerRow}{Normalize per row if TRUE, per column if FALSE}
}
\value{
Normalized object
}
\description{
This function performs Z-score normalisation on input. Dataframe input will be processed per row.
}
\examples{
Znorm(iris[,-5])
}
