% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDEA.R
\name{KDEA}
\alias{KDEA}
\title{KDEA}
\usage{
KDEA(
  dataset = NA,
  f_dataset_class_column_id = NA,
  s_PhenoDataFrame = NA,
  s_omitNA = TRUE,
  s_partitionlength = 0.5,
  s_k = floor(sqrt(dim(dataset)[1])),
  s_pvalTH = 0.05,
  s_AmountSignTH = floor(s_k * 0.3),
  s_logFCTH = NA,
  s_CovFormula = NA,
  s_CovOfImportance = NA,
  s_MakePlot = TRUE,
  s_Verbose = TRUE,
  s_maxCPUCores = 100
)
}
\arguments{
\item{dataset}{Input the dataset object (row=sample; col=feature; 'class' column required).}

\item{f_dataset_class_column_id}{Index number of class column (eg. for iris dataset classcolumn ID is 5 (species)). Can be added to s_PhenoDataFrame.}

\item{s_PhenoDataFrame}{This is a dataframe containing the phenotypic observations, rows as samples, columns as features. Make sure one of columns has exact identifiers found in rownames(dataset)}

\item{s_omitNA}{Remove NAs from data.}

\item{s_partitionlength}{Percentage number used in resampling (0-1) (default 0.5).}

\item{s_k}{Amount of folds (auto generated based on sqrt(sampleNo).}

\item{s_pvalTH}{Significance threshold.}

\item{s_AmountSignTH}{Amount of times found to be significant throughout all folds threshold (integer).}

\item{s_logFCTH}{Minimum logFC required threshold (float).}

\item{s_CovFormula}{Defaults to "~Class"; can be changed, however, the column entered in f_dataset_class_column_id is called class, and in formula should be treated so. Example: in iris the function can be run using KDEA(iris,5). The fifth column ("species") will be renamed to "Class" therefore the formula should be "~Class" and not "~Species". Other columns keep their names, therefore "~Class+Sepal.Width" is possible. This function extracts the 1st term from the formula, this is set by s_CovOfImportance, which can be changed appropriately.}

\item{s_CovOfImportance}{Automatically searches the correct "Class" column, usally 1; This is the column which is extracted after running limma::topTable. When s_CovFormula = NA then the first column is the LogFC column. When edited to eg "~Class+Sepal.Width", the first column is the "Class" FC and the second column is the effect of "Sepal.Width". Make sure this matches.}

\item{s_MakePlot}{Create plots (can take a long time when evaluating big datasets (50k+ features) set to TRUE (default)}

\item{s_Verbose}{Show more info if TRUE (default)}
}
\value{
A list of imporant features, a dataframe of ranked features, the plot, and indermediate results.
}
\description{
Leave-p-Out Cross-Validation Differential Expression Analysis (KDEA or LpO CV DEA)
Purpose:
Repeated random-sampled Limma analysis is applied to get estimates of variability in DEA, then by selecting the most robust features,
improved follow-up analysis is possible. Please see details for required dataset structure!
}
\details{
Dataset:
Requires a dataset with rows as samples and columns as features.
Rownames should contain the unique identifier of the samples.
It should contain a "Class" column to be able to detect classes. (OR use phenotypic data Classes)

Phenotypic data:
Can be added to account for covariates using formula.
Requires rows as samples and columns as features.
Should contain a column with unique identifiers of the samples.
It should contain a "Class" column to be able to detect classes. (OR use dataset Classes)
}
\examples{
KDEA(dataset = dataset, s_PhenoDataFrame = Phenotipic_data, s_CovFormula = "~Class + age + sex")
KDEA(dataset = mtcars, f_dataset_class_column_id = which(colnames(mtcars)=="am"), s_CovFormula = "~Class + gear")
KDEA(dataset = iris, f_dataset_class_column_id = 5, s_CovOfImportance = 1)
}
