% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCC.R
\name{MCC}
\alias{MCC}
\title{MCC}
\usage{
MCC(data, lev = NULL, model = NULL, showCM = FALSE)
}
\arguments{
\item{data}{Input the dataset object (column 1 = OBS; column 2 = PRED).}

\item{lev}{NULL}

\item{model}{NULL}

\item{showCM}{Show output of confusion matrix & metrics as cat.}
}
\value{
Several metrics to indicate machine learning performance ("MCC","Accuracy","Recall","Precision","F1","CombinedScore")
}
\description{
This function required a dataset with column 1 as observations and column 2 with predictions. Two-class only!
}
\examples{
MCC(data.frame(obs=c(1,1,1,2,2,2),pred=c(1,1,1,1,2,2)))
}
